Address @LysandreJik's comments
